relocated isProductinDatabase from validation.js to product.js in services-folder
